agora é possível alterar a senha

+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\User;
 
 class UserController extends Controller
@@ -12,13 +13,30 @@ class UserController extends Controller
 	}
 	
 	public function index(){
+		$user = Auth::user();
+		if(!$user->hasPermissionGlobalbyPerfil([1])){
+			return redirect("/");
+		}
+
 		$users = User::all();
 		return view("usuario.index",compact("users"));
 	}
 	public function new(){
+		$user = Auth::user();
+		if(!$user->hasPermissionGlobalbyPerfil([1])){
+			return redirect("/");
+		}
+
+
 		return view("usuario.new");
 	}
 	public function newPost(Request $request){
+		$user = Auth::user();
+		if(!$user->hasPermissionGlobalbyPerfil([1])){
+			return redirect("/");
+		}
+
+		
 		$requisicao = $request->all();
 		$validator = \Validator::make($requisicao, [
 			'name' => 'required|string|max:255',
@@ -36,10 +54,21 @@ class UserController extends Controller
 		return redirect("/usuario/edit/".$user->id);
 	}
 	public function edit($id){
+		if(!$user->hasPermissionGlobalbyPerfil([1])){
+			return redirect("/");
+		}
+
+		
 		$user = User::find($id);
 		return view("usuario.edit",compact("user"));
 	}
 	public function editPost(Request $request, $id){
+		$user = Auth::user();
+		if(!$user->hasPermissionGlobalbyPerfil([1])){
+			return redirect("/");
+		}
+
+		
 		$requisicao = $request->all();
 		$user = User::find($id);
 		if($user->email != $request->input("email")){
@@ -63,11 +92,21 @@ class UserController extends Controller
 		$user->save();
 		return redirect("/usuario/edit/".$user->id);
 	}
-	public function password($id){
+	public function password($id, Request $request){
+		$USER = Auth::user();
+		if(!$USER->hasPermissionGlobalbyPerfil([1]) && $USER->id != $id){
+			return redirect("/");
+		}
 		$user = User::find($id);
-		return view("usuario.password",compact("user"));
+		$ok = 0;
+		if($request->has("ok")) $ok = $request->input("ok");
+		return view("usuario.password",compact("user","ok","USER"));
 	}
 	public function passwordPost(Request $request, $id){
+		$USER = Auth::user();
+		if(!$USER->hasPermissionGlobalbyPerfil([1]) && $USER->id != $id){
+			return redirect("/");
+		}
 		$requisicao = $request->all();
 		$validator = \Validator::make($requisicao, [
 			'password' => 'required|string|min:6|confirmed',
@@ -78,9 +117,20 @@ class UserController extends Controller
 		$user = User::find($id);
 		$user->password = bcrypt($request->input('password'));
 		$user->save();
+		
+		if(!$USER->hasPermissionGlobalbyPerfil([1])){
+			return redirect("/usuario/password/".$user->id."?ok=1");
+		}
+
 		return redirect("/usuario/edit/".$user->id);
 	}
 	public function delete($id){
+		$user = Auth::user();
+		if(!$user->hasPermissionGlobalbyPerfil([1])){
+			return redirect("/");
+		}
+
+		
 		$user = User::find($id);
 		$user->delete();
 		return redirect("/usuario");
